+++ app/src/main/java/com/codesoom/assignment/application/AuthenticationService.java
@@ -1,51 +1,30 @@
 package com.codesoom.assignment.application;
 
-import com.codesoom.assignment.domain.Role;
-import com.codesoom.assignment.domain.RoleRepository;
-import com.codesoom.assignment.domain.User;
-import com.codesoom.assignment.domain.UserRepository;
-import com.codesoom.assignment.errors.LoginFailException;
+import com.codesoom.assignment.security.UserAuthentication;
 import com.codesoom.assignment.utils.JwtUtil;
 import io.jsonwebtoken.Claims;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
+/**
+ * 인증을 관리하는 서비스
+ */
 @Service
 public class AuthenticationService {
-    private final UserRepository userRepository;
-    private final RoleRepository roleRepository;
+    private final RoleService roleService;
     private final JwtUtil jwtUtil;
-    private final PasswordEncoder passwordEncoder;
 
-    public AuthenticationService(UserRepository userRepository,
-                                 RoleRepository roleRepository,
-                                 JwtUtil jwtUtil,
-                                 PasswordEncoder passwordEncoder) {
-        this.userRepository = userRepository;
-        this.roleRepository = roleRepository;
+    public AuthenticationService(RoleService roleService, JwtUtil jwtUtil) {
+        this.roleService = roleService;
         this.jwtUtil = jwtUtil;
-        this.passwordEncoder = passwordEncoder;
     }
 
-    public String login(String email, String password) {
-        User user = userRepository.findByEmail(email)
-                .orElseThrow(() -> new LoginFailException(email));
-
-        if (!user.authenticate(password, passwordEncoder)) {
-            throw new LoginFailException(email);
-        }
-
-        return jwtUtil.encode(user.getId());
+    public UserAuthentication authenticate(String accessToken) {
+        Long userId = getUserId(accessToken);
+        return new UserAuthentication(userId, roleService.getRoles(userId));
     }
 
-    public Long parseToken(String accessToken) {
+    public Long getUserId(String accessToken) {
         Claims claims = jwtUtil.decode(accessToken);
         return claims.get("userId", Long.class);
-    }
-
-    public List<Role> roles(Long userId) {
-        return roleRepository.findAllByUserId(userId);
     }
 }
