+++ app/src/main/java/com/codesoom/assignment/application/SessionService.java
@@ -1,0 +1,37 @@
+package com.codesoom.assignment.application;
+
+import com.codesoom.assignment.domain.User;
+import com.codesoom.assignment.domain.UserRepository;
+import com.codesoom.assignment.errors.LoginFailException;
+import com.codesoom.assignment.utils.JwtUtil;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+
+/**
+ * session 을 관리하는 서비스
+ */
+@Service
+public class SessionService {
+    private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
+    private final JwtUtil jwtUtil;
+
+
+    public SessionService(UserRepository userRepository, PasswordEncoder passwordEncoder, JwtUtil jwtUtil) {
+        this.userRepository = userRepository;
+        this.passwordEncoder = passwordEncoder;
+        this.jwtUtil = jwtUtil;
+    }
+
+    public String login(String email, String password) {
+        User user = userRepository.findByEmail(email)
+                .orElseThrow(() -> new LoginFailException(email));
+
+        if (!user.authenticate(password, passwordEncoder)) {
+            throw new LoginFailException(email);
+        }
+
+        return jwtUtil.encode(user.getId());
+    }
+
+}
