+++ app/src/main/java/com/codesoom/assignment/controllers/SessionController.java
@@ -1,18 +1,22 @@
 package com.codesoom.assignment.controllers;
 
-import com.codesoom.assignment.application.AuthenticationService;
+import com.codesoom.assignment.application.SessionService;
 import com.codesoom.assignment.dto.SessionRequestData;
 import com.codesoom.assignment.dto.SessionResponseData;
 import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/session")
 public class SessionController {
-    private AuthenticationService authenticationService;
+    private final SessionService sessionService;
 
-    public SessionController(AuthenticationService authenticationService) {
-        this.authenticationService = authenticationService;
+    public SessionController(SessionService sessionService) {
+        this.sessionService = sessionService;
     }
 
     @PostMapping
@@ -23,7 +27,7 @@ public class SessionController {
         String email = sessionRequestData.getEmail();
         String password = sessionRequestData.getPassword();
 
-        String accessToken = authenticationService.login(email, password);
+        String accessToken = sessionService.login(email, password);
 
         return SessionResponseData.builder()
                 .accessToken(accessToken)
