+++ app/src/main/java/com/codesoom/assignment/filters/JwtAuthenticationFilter.java
@@ -1,8 +1,6 @@
 package com.codesoom.assignment.filters;
 
 import com.codesoom.assignment.application.AuthenticationService;
-import com.codesoom.assignment.domain.Role;
-import com.codesoom.assignment.security.UserAuthentication;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContext;
@@ -14,7 +12,6 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.List;
 
 public class JwtAuthenticationFilter extends BasicAuthenticationFilter {
     private final AuthenticationService authenticationService;
@@ -35,11 +32,7 @@ public class JwtAuthenticationFilter extends BasicAuthenticationFilter {
 
         if (authorization != null) {
             String accessToken = authorization.substring("Bearer ".length());
-            Long userId = authenticationService.parseToken(accessToken);
-            List<Role> roles = authenticationService.roles(userId);
-            Authentication authentication =
-                    new UserAuthentication(userId, roles);
-
+            Authentication authentication = authenticationService.authenticate(accessToken);
             SecurityContext context = SecurityContextHolder.getContext();
             context.setAuthentication(authentication);
         }
